separate publish_group function
