+++ publisher/publish_test.py
@@ -1,6 +1,8 @@
 from do_import import *
 from import_utils import *
+import math
 
+from decimal import Decimal
 from model_import_actions import model_import_actions
 
 def test(engine,job_dir = None):
